+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2020-XX-XX
+### Added
+- features.codes: `get_nth_occurrence`, `tag_had_code`
+- features.windows: `calculate_days`, `build_windows`
+
 ## [0.1.0] - 2020-04-07
 ### Added
 - Initial release.